query_assistant 2/requirements.txt
psycopg2-binary>=2.9.9
streamlit>=1.35.0
python-dotenv>=1.0.0
pandas>=2.2.0
sqlparse>=0.5.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.1